refactor(headless): split Popover into per-file parts for tree-shaking
