+++ packages/headless/src/primitives/popover/parts.ts
@@ -1,0 +1,9 @@
+export { type PopoverProps, PopoverRoot as Root } from './popover-root';
+export { type PopoverTriggerProps, PopoverTrigger as Trigger } from './popover-trigger';
+export { type PopoverPortalProps, PopoverPortal as Portal } from './popover-portal';
+export { type PopoverPositionerProps, PopoverPositioner as Positioner } from './popover-positioner';
+export { type PopoverPopupProps, PopoverPopup as Popup } from './popover-popup';
+export { type PopoverArrowProps, PopoverArrow as Arrow } from './popover-arrow';
+export { type PopoverTitleProps, PopoverTitle as Title } from './popover-title';
+export { type PopoverDescriptionProps, PopoverDescription as Description } from './popover-description';
+export { type PopoverCloseProps, PopoverClose as Close } from './popover-close';
