+++ packages/headless/src/primitives/popover/popover-close.tsx
@@ -1,0 +1,25 @@
+'use client';
+
+import { type ComponentProps, mergeProps, renderElement } from '../../utils/render-element';
+import { usePopoverContext } from './popover-context';
+
+export interface PopoverCloseProps extends ComponentProps<'button'> {}
+
+export function PopoverClose(props: PopoverCloseProps) {
+  const { render, ...otherProps } = props;
+  const { setOpen } = usePopoverContext();
+
+  const defaultProps = {
+    type: 'button' as const,
+    'data-cl-slot': 'popover-close',
+    onClick() {
+      setOpen(false);
+    },
+  };
+
+  return renderElement({
+    defaultTagName: 'button',
+    render,
+    props: mergeProps<'button'>(defaultProps, otherProps),
+  });
+}
