+++ packages/headless/src/primitives/popover/popover-context.ts
@@ -1,0 +1,39 @@
+import type {
+  ExtendedRefs,
+  FloatingContext,
+  Placement,
+  ReferenceType,
+  UseInteractionsReturn,
+} from '@floating-ui/react';
+import { type CSSProperties, createContext, useContext } from 'react';
+import type { TransitionProps } from '../../hooks/use-transition';
+
+export interface PopoverContextValue {
+  open: boolean;
+  setOpen: (open: boolean) => void;
+  floatingContext: FloatingContext;
+  refs: ExtendedRefs<ReferenceType>;
+  floatingStyles: CSSProperties;
+  placement: Placement;
+  getReferenceProps: UseInteractionsReturn['getReferenceProps'];
+  getFloatingProps: UseInteractionsReturn['getFloatingProps'];
+  popupRef: React.RefObject<HTMLDivElement | null>;
+  arrowRef: React.MutableRefObject<SVGSVGElement | null>;
+  modal: boolean;
+  labelId: string | undefined;
+  descriptionId: string | undefined;
+  setLabelId: React.Dispatch<React.SetStateAction<string | undefined>>;
+  setDescriptionId: React.Dispatch<React.SetStateAction<string | undefined>>;
+  mounted: boolean;
+  transitionProps: TransitionProps;
+}
+
+export const PopoverContext = createContext<PopoverContextValue | null>(null);
+
+export function usePopoverContext() {
+  const ctx = useContext(PopoverContext);
+  if (!ctx) {
+    throw new Error('Popover compound components must be used within <Popover>');
+  }
+  return ctx;
+}
