+++ packages/headless/src/primitives/popover/popover-root.tsx
@@ -1,0 +1,139 @@
+'use client';
+
+import {
+  arrow,
+  autoUpdate,
+  FloatingNode,
+  FloatingTree,
+  flip,
+  offset,
+  type Placement,
+  shift,
+  useClick,
+  useDismiss,
+  useFloating,
+  useFloatingNodeId,
+  useFloatingParentNodeId,
+  useInteractions,
+  useRole,
+} from '@floating-ui/react';
+import { type ReactNode, useMemo, useRef, useState } from 'react';
+import { useControllableState } from '../../hooks/use-controllable-state';
+import { useTransition } from '../../hooks/use-transition';
+import { cssVars } from '../../utils/css-vars';
+import { PopoverContext, type PopoverContextValue } from './popover-context';
+
+export interface PopoverProps {
+  open?: boolean;
+  defaultOpen?: boolean;
+  onOpenChange?: (open: boolean) => void;
+  placement?: Placement;
+  sideOffset?: number;
+  modal?: boolean;
+  children: ReactNode;
+}
+
+function PopoverInner(props: PopoverProps) {
+  const nodeId = useFloatingNodeId();
+  const { placement: placementProp = 'bottom', sideOffset = 4, modal = false, children } = props;
+
+  const [open, setOpen] = useControllableState(props.open, props.defaultOpen ?? false, props.onOpenChange);
+
+  const [labelId, setLabelId] = useState<string | undefined>();
+  const [descriptionId, setDescriptionId] = useState<string | undefined>();
+
+  const arrowRef = useRef<SVGSVGElement | null>(null);
+  const popupRef = useRef<HTMLDivElement | null>(null);
+
+  const {
+    refs,
+    floatingStyles,
+    context: floatingContext,
+    placement,
+  } = useFloating({
+    nodeId,
+    open,
+    onOpenChange: setOpen,
+    placement: placementProp,
+    middleware: [
+      offset(sideOffset),
+      flip({
+        crossAxis: placementProp.includes('-'),
+        fallbackAxisSideDirection: 'end',
+        padding: 5,
+      }),
+      shift({ padding: 5 }),
+      arrow({ element: arrowRef }),
+      cssVars({ sideOffset }),
+    ],
+    whileElementsMounted: autoUpdate,
+  });
+
+  const { mounted, transitionProps } = useTransition({
+    open,
+    ref: popupRef,
+  });
+
+  const click = useClick(floatingContext);
+  const dismiss = useDismiss(floatingContext);
+  const role = useRole(floatingContext);
+
+  const { getReferenceProps, getFloatingProps } = useInteractions([click, dismiss, role]);
+
+  const contextValue = useMemo<PopoverContextValue>(
+    () => ({
+      open,
+      setOpen,
+      floatingContext,
+      refs,
+      floatingStyles,
+      placement,
+      getReferenceProps,
+      getFloatingProps,
+      popupRef,
+      arrowRef,
+      modal,
+      labelId,
+      descriptionId,
+      setLabelId,
+      setDescriptionId,
+      mounted,
+      transitionProps,
+    }),
+    [
+      open,
+      setOpen,
+      floatingContext,
+      refs,
+      floatingStyles,
+      placement,
+      getReferenceProps,
+      getFloatingProps,
+      modal,
+      labelId,
+      descriptionId,
+      mounted,
+      transitionProps,
+    ],
+  );
+
+  return (
+    <FloatingNode id={nodeId}>
+      <PopoverContext.Provider value={contextValue}>{children}</PopoverContext.Provider>
+    </FloatingNode>
+  );
+}
+
+export function PopoverRoot(props: PopoverProps) {
+  const parentId = useFloatingParentNodeId();
+
+  if (parentId === null) {
+    return (
+      <FloatingTree>
+        <PopoverInner {...props} />
+      </FloatingTree>
+    );
+  }
+
+  return <PopoverInner {...props} />;
+}
