+++ packages/headless/src/primitives/popover/popover-trigger.tsx
@@ -1,0 +1,31 @@
+'use client';
+
+import React from 'react';
+import { type ComponentProps, mergeProps, renderElement } from '../../utils/render-element';
+import { usePopoverContext } from './popover-context';
+
+export interface PopoverTriggerProps extends ComponentProps<'button'> {}
+
+export function PopoverTrigger(props: PopoverTriggerProps) {
+  const { render, ...otherProps } = props;
+  const { open, refs, getReferenceProps } = usePopoverContext();
+
+  const state = { open };
+
+  const defaultProps = {
+    type: 'button' as const,
+    'data-cl-slot': 'popover-trigger',
+    ref: refs.setReference,
+    ...(getReferenceProps() as React.ComponentPropsWithRef<'button'>),
+  };
+
+  return renderElement({
+    defaultTagName: 'button',
+    render,
+    state,
+    stateAttributesMapping: {
+      open: (v: boolean): Record<string, string> | null => (v ? { 'data-cl-open': '' } : { 'data-cl-closed': '' }),
+    },
+    props: mergeProps<'button'>(defaultProps, otherProps),
+  });
+}
